+++ src/cot_synthesis/generate.py
@@ -80,7 +80,7 @@ def generate_vllm(prompts: list[str], model: str, n: int, temperature: float,
                   top_p: float, max_tokens: int, tensor_parallel_size: int = 1) -> list[list[str]]:
     from vllm import LLM, SamplingParams
 
-    llm = LLM(model=model, dtype="bfloat16", gpu_memory_utilization=0.90,
+    llm = LLM(model=model, dtype="float16", gpu_memory_utilization=0.90,
               max_model_len=max(4096, max_tokens + 1024), trust_remote_code=True,
               tensor_parallel_size=tensor_parallel_size)
     sp = SamplingParams(n=n, temperature=temperature, top_p=top_p, max_tokens=max_tokens)
